Use FP32 for position and texcoords
We need that extra precision for sampling large textures

+++ app/shaders/d3d11_bt601lim_pixel.hlsl
@@ -16,8 +16,8 @@ static const min16float3 offsets =
 
 struct ShaderInput
 {
-    min16float4 pos : SV_POSITION;
-    min16float2 tex : TEXCOORD0;
+    float4 pos : SV_POSITION;
+    float2 tex : TEXCOORD0;
 };
 
 min16float4 main(ShaderInput input) : SV_TARGET
